state when revalidate for dynamic data and user
