+++ package.json
@@ -21,7 +21,8 @@
 		"nookies": "^2.5.0",
 		"react": "17.0.1",
 		"react-dom": "17.0.1",
-		"react-twitter-widgets": "^1.9.5"
+		"react-twitter-widgets": "^1.9.5",
+		"swr": "^0.4.0"
 	},
 	"devDependencies": {
 		"@types/marked": "^1.2.1",
