+++ src/components/Banner/index.tsx
@@ -1,18 +1,14 @@
-import { useState, useEffect } from 'react';
+import useSWR from 'swr';
 
 import { fetchAllBanners, IBanner } from '@/lib/dynamicData/banner';
 import Banner from './banner';
 
 export default function Banners() {
-	const [banners, setBanners] = useState<null | Array<IBanner>>(null);
+	const { data } = useSWR('all_banners', fetchAllBanners, { refreshInterval: 60 * 1000 });
 
-	useEffect(() => {
-		fetchAllBanners()
-			.then((bns) => setBanners(bns))
-			.catch(() => setBanners([]));
-	}, []);
+	let banners: IBanner[] = data ?? [];
 
-	if (Array.isArray(banners) && banners.length > 0) {
+	if (banners.length > 0) {
 		// TODO: Change the banner style
 		return (
 			<>
