+++ src/lib/user/session.ts
@@ -1,5 +1,6 @@
 import type { GetServerSidePropsContext } from 'next';
 import type { ParsedUrlQuery } from 'querystring';
+import axios from 'axios';
 
 import { setCookie } from 'nookies';
 
@@ -13,4 +14,15 @@ export function storeUserWithToken(
 			path: '/',
 		});
 	}
+}
+
+export function getSessionAuthToken(): Promise<string> {
+	return axios
+		.get('/api/session')
+		.then((response) => response.data)
+		.then((data) => data.auth_token)
+		.catch((err) => {
+			console.log(err, 'while fetching user session');
+			return null;
+		});
 }
